Module kik_sync_service::channel

DeliveryService is the channel used for the synchronous message-sharing and work. It can be created with DeliveryService::default values or be customized by using ChannelConfig as argument for DeliveryService::new.

Structs

ChannelConfig
To be used when the user needs to set specific configurations before creating a DeliveryService channel. Optional type.

DeliveryService
Main structure for the entire crate. Creates the channels, workers and feeder.