Trait kik_sync_service::message::Message

pub trait Message<T, R>: Sync + Send + Clone + 'static where
    R: MessageInput<T>,
    T: MessageData, {
    fn set_input(&mut self, message_input: R);

    fn work(&mut self);

    fn clone_message_data(&self) -> T;

    fn new() -> Self;

    fn new_message_input() -> R { ... }

    fn new_message_data() -> T { ... }
}

Message has the tools to generate each MessageData T, based on each MessageInput R. Must implement Sync, Send, Clone and have lifetime 'static.

Required methods

fn set_input(&mut self, message_input: R)

Behavior for storing a given input MessageInput, before a worker can use it for generating MessageData. Used by kik_feeder.

fn work(&mut self)

Workers will call this to use the stored MessageInput (R) to generate and replace the existing MessageData stored. Used by kik_worker.

fn clone_message_data(&self) -> T

This method is used when retrieving MessageData for the iterator. Clone the MessageData stored and return it. Used by kik_feeder.

fn new() -> Self

Construct a new message with default values. Used by kik_feeder.

Loading content...
Provided methods

fn new_message_input() -> R

This will call MessageInput::new() method. No need to implement this. Used by kik_feeder.

fn new_message_data() -> T

This will call MessageData::new() method. No need to implement this. Used by kik_feeder.

Loading content...